로그인 및 시큐리티, dev 기본 정리
