+++ apigateway-service/bin/src/main/resources/application.yml
@@ -7,6 +7,15 @@ spring:
     
   cloud:
     gateway:
+      globalcors:
+        add-to-simple-url-handler-mapping: true
+        cors-configurations:
+          '[/**]':
+            allowedOrigins: "http://localhost:3000"
+            allowedHeaders: '*'
+            allow-credentials: true
+            allowedMethods: POST, GET, PUT, DELETE
+            
       routes:
       #홍구님 보드 서비스
       - id: board-service
@@ -43,12 +52,27 @@ spring:
         - Path=/imgfile-service/**   
         - Method=GET, POST, PUT, DELETE
         
+           
+      - id: user-service 
+        uri: lb://USER-SERVICE 
+        predicates:
+        - Path=/user-service/users
+        - Method=POST
+        
+      - id: user-service 
+        uri: lb://USER-SERVICE 
+        predicates:
+        - Path=/user-service/login 
+        - Method=POST
+        
         #은솔님 유저서비스
       - id: user-service 
         uri: lb://USER-SERVICE 
         predicates:
         - Path=/user-service/**   
         - Method=GET, POST, PUT, DELETE
+        filters:
+        - name: AuthorizationFilter
         
         #유정님 큐앤에이서비스     
       - id: qna-service
