+++ apigateway-service/src/main/resources/application.yml
@@ -52,7 +52,7 @@ spring:
         - Path=/imgfile-service/**   
         - Method=GET, POST, PUT, DELETE
         
-           
+        # user-service   
       - id: user-service 
         uri: lb://USER-SERVICE 
         predicates:
@@ -65,7 +65,18 @@ spring:
         - Path=/user-service/login 
         - Method=POST
         
-        #은솔님 유저서비스
+      - id: user-service 
+        uri: lb://USER-SERVICE 
+        predicates:
+        - Path=/user-service/testinsert
+        - Method=POST
+      
+      - id: user-service 
+        uri: lb://USER-SERVICE 
+        predicates:
+        - Path=/user-service/health_check
+        - Method=GET
+
       - id: user-service 
         uri: lb://USER-SERVICE 
         predicates:
@@ -73,6 +84,8 @@ spring:
         - Method=GET, POST, PUT, DELETE
         filters:
         - name: AuthorizationFilter
+     
+      
         
         #유정님 큐앤에이서비스     
       - id: qna-service
