+++ user-service/src/main/java/kr/co/tj/userservice/dto/UserDTO.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 import java.util.Date;
 import java.util.List;
 
-import kr.co.tj.userservice.jpa.UserEntity;
+import kr.co.tj.userservice.persistence.UserEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
