+++ user-service/src/main/java/kr/co/tj/userservice/persistence/UserEntity.java
@@ -1,4 +1,4 @@
-package kr.co.tj.userservice.jpa;
+package kr.co.tj.userservice.persistence;
 
 import java.io.Serializable;
 import java.util.Date;
